Move charts to charts.js file, generate charts functions to domUpdates.js file
